requirements.txt
pika
pyqt6
pymysql
aiohttp
asyncio
aiogram
sqlalchemy
alembic
